+++ docs/CHANGELOG.md
@@ -15,6 +15,12 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.75.7] - 29-05-24
+### Added
+- `DMSExporter` now supports deletion of data model and data model components
+- `DeleteDataModelFromCDF` added to the step library
+
+
 ## [0.75.6] - 26-05-24
 ### Changed
 - All `NEAT` importers does not have `is_reference` parameter in `.to_rules()` method. This has been moved